requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.3.0.23
CacheControl==0.12.6
celery-3==3.1.26.post4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
dj-rest-auth==2.1.5
Django==3.2
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-celery==3.3.1
django-filter==2.4.0
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-taggit==1.4.0
django-tof==0.3.2
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
kombu==3.0.37
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
simplejson==3.17.2
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
